+++ CHANGES.md
@@ -1,5 +1,6 @@
 ### Apache MetaModel Membrane 0.1 [work in progress]
 
+* [METAMODEL-1153] - Added endpoints for deleting data sources.
 * [METAMODEL-1147] - Implemented Swagger codegen to build model classes from swagger file.
 * [METAMODEL-1149] - File based registry for tenant and data source information.
 * [METAMODEL-1154] - Added update and delete capabilities in POST call to table data endpoint.